+++ ddtrace/contrib/util.py
@@ -1,33 +1,16 @@
-from importlib import import_module
+# [Backward compatibility]: keep importing modules functions
+from ..utils.deprecation import deprecation
+from ..utils.importlib import require_modules, func_name, module_name
 
 
-class require_modules(object):
-    """
-    Context manager to check the availability of required modules.
-    """
-    def __init__(self, modules):
-        self._missing_modules = []
-        for module in modules:
-            try:
-                import_module(module)
-            except ImportError:
-                self._missing_modules.append(module)
+deprecation(
+    name='ddtrace.contrib.util',
+    message='Use `ddtrace.utils.importlib` module instead',
+    version='1.0.0',
+)
 
-    def __enter__(self):
-        return self._missing_modules
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        return False
-
-
-def func_name(f):
-    """
-    Return a human readable version of the function's name.
-    """
-    if hasattr(f, '__module__'):
-        return "%s.%s" % (f.__module__, getattr(f, '__name__', f.__class__.__name__))
-    return getattr(f, '__name__', f.__class__.__name__)
-
-
-def module_name(instance):
-    return instance.__class__.__module__.split('.')[0]
+__all__ = [
+    'require_modules',
+    'func_name',
+    'module_name',
+]
